requirements.txt
flask
python-dotenv
flask-smorest
pyodbc
flask-jwt-extended
bson
pymongo
python-dotenv
